Restructure main body and redefine bounding space
Created a new `div` right between `body` and the rest of the content, allowing me to define an initial body height. Before doing that I had to redefine the bounding space by converting the left and right margins into a general margin with `5vh` on every edge. From there I can unify the units used with the container, using `82.4vh`, this creates an initial body height with room for it to expand vertically